requirements.txt
appnope==0.1.3
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==4.4.2
distlib==0.3.8
executing==1.2.0
filelock==3.13.1
fonttools==4.39.0
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
moviepy==1.0.3
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
proglog==0.1.10
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==25.0.2
requests==2.30.0
scipy==1.11.4
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
urllib3==2.0.2
virtualenv==20.25.0
wcwidth==0.2.6
